requirements.txt
requests~=2.31
pandas~=2.1
psycopg2-binary~=2.9
prefect~=2.15
